# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Moustashari AI - Agentic AI-powered business consulting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List consulting services
      parameters:
        - name: categoryId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: minPrice
          in: query
          schema:
            type: ["number", "null"]
        - name: maxPrice
          in: query
          schema:
            type: ["number", "null"]
      responses:
        "200":
          description: List of services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /services/{id}:
    get:
      operationId: getService
      tags: [services]
      summary: Get service detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Service detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ServiceDetail"
        "404":
          description: Not found
...
components:
  schemas:
...
    Service:
      type: object
      required: [id, title, description, price, duration, categoryId, categoryName, consultantId, consultantName, rating, reviewCount, tags, imageUrl]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        price:
          type: number
        duration:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        consultantId:
          type: integer
        consultantName:
          type: string
        consultantAvatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
    ServiceDetail:
      type: object
      required: [id, title, description, price, duration, categoryId, categoryName, consultant, rating, reviewCount, tags, deliverables, faqs]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        longDescription:
          type: ["string", "null"]
        price:
          type: number
        duration:
          type: string
        categoryId:
          type: integer
        categoryName:
          type: string
        consultant:
          $ref: "#/components/schemas/Consultant"
        rating:
          type: number
        reviewCount:
          type: integer
        tags:
          type: array
          items:
            type: string
        imageUrl:
          type: ["string", "null"]
        featured:
          type: boolean
        deliverables:
          type: array
          items:
            type: string
        faqs:
          type: array
          items:
            $ref: "#/components/schemas/FAQ"
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
    FAQ:
      type: object
      required: [question, answer]
      properties:
        question:
          type: string
        answer:
          type: string
    Review:
      type: object
      required: [id, authorName, rating, comment, createdAt]
      properties:
        id:
          type: integer
        authorName:
          type: string
        rating:
          type: number
        comment:
          type: string
        createdAt:
          type: string

    # Consultants
    Consultant:
      type: object
      required: [id, name, title, bio, avatarUrl, rating, reviewCount, specializations, yearsExperience, serviceCount]
      properties:
        id:
          type: integer
        name:
          type: string
        title:
          type: string
        bio:
          type: string
        avatarUrl:
          type: ["string", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        specializations:
          type: array
          items:
            type: string
        yearsExperience:
          type: integer
        serviceCount:
          type: integer
        verified:
          type: boolean
...